project/requirements.txt
langchain-core>=0.2.33,<0.3
langchain-openai>=0.1.20,<0.2
langchain-community>=0.2.12,<0.3
chromadb>=0.5.6,<0.6
pydantic>=2.5,<3
python-dotenv>=1.0,<2
openai>=1.41.0,<2
tiktoken>=0.7.0,<1
# Offline version
scikit-learn>=1.3,<1.6